als_backend/requirements.txt
fastapi>=0.100.0,<1.0.0
uvicorn[standard]>=0.20.0,<1.0.0
groq>=0.4.0
pydantic>=2.0.0,<3.0.0
python-dotenv>=1.0.0
filelock>=3.0.0
numpy>=1.24.0
scipy>=1.10.0
gunicorn>=21.0.0
uvicorn[standard]>=0.20.0,<1.0.0
